:hammer: code refactoring
